delete custom decorator and move its code into a STYLING_PHASE message decorator event
